- Correct implementation of document.getScrollTop/left.
For old browsers (<= ie10), we keep the old logic and relies on a
quirk/strict mode check to find the document scroll element.
For other browsers, we check first of document.scrollingElement exists
and use it computing scrollTop/Left. If not we use document.body for
webkit browser or rely on the quirk/strict mode check for other
browsers.
We default to document.documentElement if all the logic above return
null (SVG document)

- Fix getAbsoluteLeft/Top computation for webkit browsers to rely on the
correct implementation of document.getScrollLeft/Top.

Bug: #9542
Bug-Link: https://github.com/gwtproject/gwt/issues/9542
Change-Id: I02fe4de00a3f646687f0550e603fc7bb4aca1b80